refactor: 번개모각코 생성 지도 부분 MGCMap 공통 컴포넌트로 변경
